refactor: changed some directions, feat_profile renamed to feat-profile
