+++ app/src/main/res/layout/activity_main.xml
@@ -2,7 +2,7 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    tools:context=".main.view.MainActivity"
+    tools:context=".main.MainActivity"
     android:layout_height="match_parent">
 
     <FrameLayout
