+++ settings.gradle
@@ -4,4 +4,4 @@ include ':core'
 include ':model'
 include ':utils'
 include ':repository'
-include ':feat_profile'
+include ':feat-profile'
